Add files via upload
This is the notebook which goes with the laboratory for FTIR data. There is one notebook and three data files.  The QV is for the Virgin filament material (before 3D printing), and the other three are ASTM 3D printed dogbones with a hex pattern and 30% fill. The material was Polycarbonate.