Add files via upload
This is a notebook to model heat impulse in different materials.  In this lab, students first run this program to model the heat flow and then will program an Arduino and TMPs to measure a heat impulse in bars of differing materials. They will run this notebook and then input the materials they plan to test experimentally. The next notebook will show them how to use the experimental data to extract out the conductivity and diffusivity of the materials and compare to the theoretical values and peaks from this notebook.